udah ga dobel lagi

+++ image_upload.php
@@ -12,30 +12,29 @@ if(isset($_POST['submit'])) {
 
         
         if(in_array($ext, $extension)) {
-            $filename = str_replace('.', '_', basename($filename, $ext)) ;
-            $newfilename = $filename.time().".".$ext;
+            // $filename = str_replace('.', '_', basename($filename, $ext)) ;
+            // $newfilename = $filename.time().".".$ext;
             $path = 'public/property/'.$name;
             if (!file_exists($path)) {
                 mkdir($path, 0777, true);
             }
-            $image = 'public/property/'.$name.'/'.$newfilename;
+            // $image = 'public/property/'.$name.'/'.$newfilename;
+            // move_uploaded_file($tmp_filename, $image);
+            // $array_image[] = $image;
+
+            if(!file_exists('public/property/'.$name.'/'.$filename)) {
+            $image = 'public/property/'.$name.'/'.$filename;
             move_uploaded_file($tmp_filename, $image);
             $array_image[] = $image;
 
-            // if(!file_exists('images/'.$filename)) {
-            // move_uploaded_file($tmp_filename, 'images/'.$filename);
-            // $image = $filename;
-            // } else {
-            //     $filename = str_replace('.', '_', basename($filename, $ext)) ;
-            //     $newfilename = $filename.time().".".$ext;
-            //     move_uploaded_file($tmp_filename, 'images/'.$newfilename);
-            //     $image = $newfilename;
-            // }
+            } else {
+                $filename = str_replace('.', '_', basename($filename, $ext)) ;
+                $newfilename = $filename.time().".".$ext;
+                $image = 'public/property/'.$name.'/'.$newfilename;
+                move_uploaded_file($tmp_filename, 'images/'.$newfilename);
+                $array_image[] = $image;
+            }
 
-            // $insert_img = "INSERT INTO `property_list`(`images`) VALUES ('$image')";
-            // mysqli_query($con, $insert_img);
-
-            // header("Location : index");
         } 
     }
